fix: make Flow work nicely
You can now have Flow types without a separate build step.
