+++ .gitignore
@@ -3,6 +3,7 @@ package-lock.json
 **/node_modules/
 **/*.log
 test/repo-tests*
+yarn.lock
 
 # Logs
 logs
