+++ package.json
@@ -3,24 +3,21 @@
   "version": "0.4.13",
   "description": "multihash implementation",
   "lead-maintainer": "David Dias <daviddias@ipfs.io>",
-  "main": "lib/index.js",
+  "main": "dist/src/index.js",
   "scripts": {
     "lint": "aegir lint && npm run type-check",
-    "test:browser": "npm run build && aegir test --target browser",
-    "test:node": "npm run build:node && aegir test --target node",
+    "test:browser": "aegir test --target browser",
+    "test:node": "aegir test --target node",
     "type-check": "flow check",
-    "build": "npm run build:node && BABEL_ENV=umd aegir build",
-    "test": "npm run build:node && aegir test",
+    "build": "flow-remove-types src/ --out-dir dist/src",
+    "test": "aegir test --eek",
     "docs": "aegir docs",
     "release": "aegir release --docs",
     "release-minor": "aegir release --type minor --docs",
     "release-major": "aegir release --type major --docs",
     "coverage": "aegir coverage",
     "coverage-publish": "aegir coverage --provider coveralls",
-    "build:types": "flow-copy-source --verbose src lib",
-    "build:lib": "babel --out-dir lib src",
-    "build:node": "npm run build:types && npm run build:lib",
-    "start": "flow-copy-source --watch --verbose src lib & babel --watch --out-dir lib src"
+    "build:types": "flow-copy-source --verbose src lib"
   },
   "pre-commit": [
     "lint",
@@ -44,6 +41,8 @@
   },
   "dependencies": {
     "bs58": "^4.0.1",
+    "esm": "^3.0.36",
+    "flow-remove-types": "^1.2.3",
     "varint": "^5.0.0"
   },
   "devDependencies": {
