+++ src/index.js
@@ -1,233 +1,5 @@
-// @flow
-
-/**
- * Multihash implementation in JavaScript.
- *
- * @module multihash
- */
-import bs58 from 'bs58'
-import varint from 'varint'
-import {names, codes, defaultLengths, type Name, type Code} from './constants'
-
-export {names, codes, defaultLengths}
-
-export type Multihash = Buffer
-export type MultihashPrefix = Buffer
-export type B58Buffer = Buffer
-export type HexString = string
-export type MultihashData = {
-  code: Code,
-  name: Name,
-  length: number,
-  digest: Buffer
+if(!process.env.IS_WEBPACK_BUILD){
+    require('flow-remove-types/register')
+    require = require("esm")(module/*, options*/)
 }
-
-/**
- * Convert the given multihash to a hex encoded string.
- *
- * @param {Buffer} hash
- * @returns {string}
- */
-export function toHexString (hash: Multihash): HexString {
-  if (!Buffer.isBuffer(hash)) {
-    throw new Error('must be passed a buffer')
-  }
-
-  return hash.toString('hex')
-}
-
-/**
- * Convert the given hex encoded string to a multihash.
- *
- * @param {string} hash
- * @returns {Buffer}
- */
-export function fromHexString (hash: HexString): Multihash {
-  return Buffer.from(hash, 'hex')
-}
-
-export type B58String = string
-/**
- * Convert the given multihash to a base58 encoded string.
- *
- * @param {Buffer} hash
- * @returns {string}
- */
-export function toB58String (hash: Multihash): B58String {
-  if (!Buffer.isBuffer(hash)) {
-    throw new Error('must be passed a buffer')
-  }
-
-  return bs58.encode(hash)
-}
-
-/**
- * Convert the given base58 encoded string to a multihash.
- *
- * @param {string|Buffer} hash
- * @returns {Buffer}
- */
-export function fromB58String (hash: B58String | B58Buffer): Multihash {
-  let encoded = hash
-  if (Buffer.isBuffer(hash)) {
-    encoded = hash.toString()
-  }
-
-  return Buffer.from(bs58.decode(encoded))
-}
-
-/**
- * Decode a hash from the given multihash.
- *
- * @param {Buffer} buf
- * @returns {{code: number, name: string, length: number, digest: Buffer}} result
- */
-export function decode (buf: Multihash): MultihashData {
-  if (!(Buffer.isBuffer(buf))) {
-    throw new Error('multihash must be a Buffer')
-  }
-
-  if (buf.length < 3) {
-    throw new Error('multihash too short. must be > 3 bytes.')
-  }
-
-  let code = varint.decode(buf)
-  if (!isValidCode(code)) {
-    throw new Error(`multihash unknown function code: 0x${code.toString(16)}`)
-  }
-  buf = buf.slice(varint.decode.bytes)
-
-  let len = varint.decode(buf)
-  if (len < 1) {
-    throw new Error(`multihash invalid length: 0x${len.toString(16)}`)
-  }
-  buf = buf.slice(varint.decode.bytes)
-
-  if (buf.length !== len) {
-    throw new Error(`multihash length inconsistent: 0x${buf.toString('hex')}`)
-  }
-
-  return {
-    code: code,
-    name: codes[code],
-    length: len,
-    digest: buf
-  }
-}
-
-/**
- *  Encode a hash digest along with the specified function code.
- *
- * > **Note:** the length is derived from the length of the digest itself.
- *
- * @param {Buffer} digest
- * @param {string|number} code
- * @param {number} [length]
- * @returns {Buffer}
- */
-export function encode (digest: Buffer, code: Code, length: number): Multihash {
-  if (!digest || !code) {
-    throw new Error('multihash encode requires at least two args: digest, code')
-  }
-
-  // ensure it's a hashfunction code.
-  const hashfn = coerceCode(code)
-
-  if (!(Buffer.isBuffer(digest))) {
-    throw new Error('digest should be a Buffer')
-  }
-
-  if (length == null) {
-    length = digest.length
-  }
-
-  if (length && digest.length !== length) {
-    throw new Error('digest length should be equal to specified length.')
-  }
-
-  return Buffer.concat([
-    Buffer.from(varint.encode(hashfn)),
-    Buffer.from(varint.encode(length)),
-    digest
-  ])
-}
-
-/**
- * Converts a hash function name into the matching code.
- * If passed a number it will return the number if it's a valid code.
- * @param {string|number} name
- * @returns {number}
- */
-// TODO: Sholud we just allow Name here instead ?
-export function coerceCode (name: Name | Code): Code {
-  let code = name
-
-  if (typeof name === 'string') {
-    if (!names[name]) {
-      throw new Error(`Unrecognized hash function named: ${name}`)
-    }
-    code = names[name]
-  }
-
-  if (typeof code !== 'number') {
-    throw new Error(`Hash function code should be a number. Got: ${code}`)
-  }
-
-  if (!codes[code] && !exports.isAppCode(code)) {
-    throw new Error(`Unrecognized function code: ${code}`)
-  }
-
-  return code
-}
-
-/**
- * Checks wether a code is part of the app range
- *
- * @param {number} code
- * @returns {boolean}
- */
-export function isAppCode (code: number): boolean {
-  return code > 0 && code < 0x10
-}
-
-/**
- * Checks whether a multihash code is valid.
- *
- * @param {number} code
- * @returns {boolean}
- */
-export function isValidCode (code: number): boolean {
-  if (isAppCode(code)) {
-    return true
-  }
-
-  if (codes[(code: any)]) {
-    return true
-  }
-
-  return false
-}
-
-/**
- * Check if the given buffer is a valid multihash. Throws an error if it is not valid.
- *
- * @param {Buffer} multihash
- * @returns {undefined}
- * @throws {Error}
- */
-export function validate (multihash: Buffer) {
-  decode(multihash) // throws if bad.
-}
-
-/**
- * Returns a prefix from a valid multihash. Throws an error if it is not valid.
- *
- * @param {Buffer} multihash
- * @returns {Buffer}
- * @throws {Error}
- */
-export function prefix (multihash: Multihash): MultihashPrefix {
-  validate(multihash)
-
-  return multihash.slice(0, 2)
-}
+module.exports = require("./multihash.js")
